simplify script
Remove a lot of superfluous code and simplify the script a bit. Furthermore, do not use the MIP folder saved by the scope anymore. Just do the max prjection ousevles. A known issue is with single plane data. It will not work anymore due to the new file strcutre that is used from the bruker software. Will need the user to specify instead