fix random bug from bruker side
There was an issue with extracting metadata from the XML file. Specifically, when using three interlaced laser tracks from the Bock microscope, Prairie View wrote an invalid character into the XML file. I tried updating the XML version from 1.0 to 1.1, but that did not resolve the issue. For now, the solution is to duplicate the XML file, locate the two problematic lines of code causing the error, delete them, and use this modified “backup file” to extract the metadata. Afterward, delete the backup file to ensure the original file remains unchanged.